job-agent/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
openai
pymupdf
requests
beautifulsoup4
playwright
sentence-transformers
